+++ packages/core/src/components/layout/box/Box.tsx
@@ -1,3 +1,4 @@
+import isPropValid from "@emotion/is-prop-valid";
 import styled from "@emotion/styled";
 import {
   BackgroundProperty,
@@ -38,19 +39,15 @@ import {
   FlexProps,
   flexWrap,
   FlexWrapProps,
-  height,
   HeightProps,
   justifyContent,
   JustifyContentProps,
+  layout,
   left,
   LeftProps,
-  maxHeight,
   MaxHeightProps,
-  maxWidth,
   MaxWidthProps,
-  minHeight,
   MinHeightProps,
-  minWidth,
   MinWidthProps,
   overflow,
   OverflowProps,
@@ -61,7 +58,6 @@ import {
   TLengthStyledSystem,
   top,
   TopProps,
-  width,
   WidthProps,
   zIndex,
   ZIndexProps
@@ -148,11 +144,25 @@ export interface BoxProps extends StyledSystemProps, DivProps {
   hoverBorder?: ThemeColorField | BorderProperty<TLengthStyledSystem>;
 }
 
-const FlexBox = styled.div<
-  FlexBoxProps &
-    BoxShadowProps &
-    BackgroundProps & { themeSpacing: number; themeIndent: number }
->`
+const excludedProps = ["spacing", "indent", "width", "height", "color"];
+
+const isExcludedWebUiProp = (propName: string) =>
+  excludedProps.indexOf(propName) !== -1;
+
+const getPaddingRule = (props: InnerProps) =>
+  props.spacing || props.indent
+    ? `padding: ${numberOrZero(props.spacing) * (props.themeSpacing || 10)}px
+    ${numberOrZero(props.indent) * (props.themeIndent || 10)}px;`
+    : "";
+
+type InnerProps = FlexBoxProps &
+  BoxShadowProps &
+  BackgroundProps & { themeSpacing: number; themeIndent: number };
+
+const FlexBox = styled("div", {
+  shouldForwardProp: propName =>
+    isExcludedWebUiProp(propName) ? false : isPropValid(propName)
+})<InnerProps>`
   box-sizing: border-box;
   display: ${props => props.display || "flex"};
   ${alignItems};
@@ -172,23 +182,16 @@ const FlexBox = styled.div<
   flex-direction: ${props =>
     (props.row && "row") || props.flexDirection || "column"};
   ${flexWrap};
-  ${height};
   ${justifyContent};
-  ${minHeight};
-  ${minWidth};
-  ${maxHeight};
-  ${maxWidth};
   ${overflow};
-  padding: ${props =>
-    numberOrZero(props.spacing) * (props.themeSpacing || 10)}px
-    ${props => numberOrZero(props.indent) * (props.themeIndent || 10)}px;
-  ${position}
-  ${width};
-  ${zIndex}
-  ${left}
-  ${right}
-  ${top}
-  ${bottom}
+  ${getPaddingRule}
+  ${position};
+  ${layout};
+  ${zIndex};
+  ${left};
+  ${right};
+  ${top};
+  ${bottom};
   :hover {
     background: ${({ hoverBackground }) => hoverBackground};
     border: ${({ hoverBorder }) => hoverBorder};