+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+#### Upcoming 0.0.32
+
+* `Box` is now responsive. `width` and `height` props can take array of values.
+    * See https://styled-system.com/responsive-styles for more info.
+* `padding` CSS rule is no longer set on `Box` div element, when spacing and indent are not set.
+* Add stories for responsive `Box`.
+* Add story for aria label on `Box`.
+
 #### 0.0.31
 
 * Add first version of `@stenajs-webui/modal`